Module vectis_factory::state

Structs

WalletInfo

Constants

ADDR_PREFIX

Chain address prefix

CLAIM_FEE

Fee for DAO when claiming govec tokens.

DAO

The DAO address where the fees for new wallet goes to, also the admin of the contract. On a remote chain, this is the remote-tunnel contract

GOVEC_CLAIM_LIST

All proxy wallets that have yet to claim their govec tokens

GOVEC_MINTER

Governing token minting contract

PROXY_CODE_ID

The latest supported wallet_proxy code id stored onchain

PROXY_MULTISIG_CODE_ID

The latest default multisig code id stored onchain for the proxy

TOTAL_CREATED

The total number of wallets successfully created by the factory i.e. if creation fail, this is not incremented

WALLET_FEE

Fee for DAO when a wallet is created